Em principio o MaxFlow esta feito

+++ City.cpp
@@ -6,13 +6,13 @@ string City::getName() {return name_;}
 
 string City::getPopulation() {return population_;}
 
-float City::getDemand() {return demand_;}
+double City::getDemand() {return demand_;}
 
 unsigned int City::getId() {return id_;}
 
 void City::setCode(std::string &code) { code_ = code;}
 
-void City::setDemand(float demand) { demand_ = demand;}
+void City::setDemand(double demand) { demand_ = demand;}
 
 void City::setId(unsigned int id) { id_ = id;}
 
@@ -20,7 +20,7 @@ void City::setName(std::string &name) { name_ = name;}
 
 void City::setPopulation(string &population) { population_ = population;}
 
-City::City(string &name, string &code, unsigned int id, string population, float demand) {
+City::City(string &name, string &code, unsigned int id, string population, double demand) {
     name_ = name;
     code_ = code;
     id_ = id;
